Refactor generate report popups and create radio group component
