+++ Clients/src/presentation/components/RadioGroup/styles.ts
@@ -1,0 +1,6 @@
+export const labelStyle = (theme: any) => ({  
+  '& .MuiFormControlLabel-label': { 
+    color: "#475467",
+    fontSize: theme.typography.fontSize,
+  }  
+})
