+++ Clients/src/presentation/components/Reporting/GenerateReport/DownloadReportFrom/styles.ts
@@ -1,0 +1,16 @@
+export const styles = {
+  container: {
+    textAlign: 'center', 
+    padding: 30
+  },
+  titleText: {
+    fontSize: 16, 
+    color: "#344054", 
+    fontWeight: "bold", 
+    mb: 4
+  },
+  baseText: {
+    color: "#344054",
+    fontSize: 13,
+  }
+}
