+++ Clients/src/presentation/components/Reporting/GenerateReport/constants.ts
@@ -1,0 +1,7 @@
+export const REPORT_TYPES = [
+  "Project risks report",
+  "Compliance tracker report",
+  "Assessment tracker report",
+  "Vendors and risks report",
+  "All reports combined in one file"
+]
