+++ Clients/src/presentation/components/Reporting/GenerateReport/styles.ts
@@ -1,0 +1,21 @@
+export const styles = {
+  formContainer: {
+    padding: 12,
+    width: { xs: "100%", sm: "398px" },
+    height: '421px'
+  },
+  contentCenter: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center'
+  },
+  iconButton: {
+    position: 'absolute', 
+    right: '12px', 
+    top: '10px'
+  },
+  closeButton: {
+    width: 24, 
+    height: 24
+  }
+}
